+++ Citadel/src/utopia/citadel/database/factory/organization/MembershipFactory.scala
@@ -5,24 +5,30 @@ import utopia.citadel.database.model.organization.MembershipModel
 import utopia.flow.datastructure.immutable.{Constant, Model}
 import utopia.metropolis.model.partial.organization.MembershipData
 import utopia.metropolis.model.stored.organization.Membership
-import utopia.vault.nosql.factory.{Deprecatable, FromValidatedRowModelFactory}
+import utopia.vault.nosql.factory.{Deprecatable, FromRowFactoryWithTimestamps, FromValidatedRowModelFactory}
 
 /**
   * Used for reading organization memberships from the database
   * @author Mikko Hilpinen
   * @since 17.6.2020, v1.0
   */
-object MembershipFactory extends FromValidatedRowModelFactory[Membership] with Deprecatable
+object MembershipFactory extends FromValidatedRowModelFactory[Membership]
+	with FromRowFactoryWithTimestamps[Membership] with Deprecatable
 {
+	// ATTRIBUTES   --------------------------
+	
+	override val creationTimePropertyName = "started"
+	
+	
 	// IMPLEMENTED	--------------------------
 	
-	override val nonDeprecatedCondition = table("ended").isNull
+	override def table = Tables.organizationMembership
+	
+	override def nonDeprecatedCondition = model.nonDeprecatedCondition
 	
 	override protected def fromValidatedModel(model: Model[Constant]) = Membership(model("id").getInt,
 		MembershipData(model(this.model.organizationIdAttName).getInt, model("userId").getInt, model("creatorId").int,
-			model("started").getInstant, model("ended").instant))
-	
-	override def table = Tables.organizationMembership
+			model(creationTimePropertyName).getInstant, model("ended").instant))
 	
 	
 	// COMPUTED	-------------------------------